modifier Post et Delete

+++ BSFlixFlex/BSFlixFlex/Controllers/MyFavoritesController.cs
@@ -3,6 +3,7 @@ using BSFlixFlex.Client.Shareds.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using System.Net;
 using System.Security.Claims;
 using System.Security.Principal;
 
@@ -15,8 +16,10 @@ namespace BSFlixFlex.Controllers
     [Authorize(Policy = "CookiesOrBearer")]
     public class MyFavoritesController(IMyFavoriteService myFavoriService) : ControllerBase
     {
-
         // GET: api/<ValuesController>
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(ApiListResponse<IDiscovryCommonProperty>))]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpGet]
         public async Task<ActionResult<ApiListResponse<IDiscovryCommonProperty>>> Get([FromQuery] int? page, [FromQuery] int? pageSize)
         {
@@ -25,36 +28,69 @@ namespace BSFlixFlex.Controllers
             {
                 return Ok(r);
             }
-            return NotFound();
+            return Problem(statusCode: 500);
         }
 
         // GET api/<ValuesController>/5
-        [HttpGet("{cinematography}")]
-        public async Task<bool> Get(Cinematography cinematography, int id)
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(bool))]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [HttpGet("{cinematography}/{id:int}")]
+        public async Task<bool> Get([FromRoute]Cinematography cinematography, [FromRoute] int id)
         {
             return await myFavoriService.IsFavoriteAsync(id, cinematography, User);
         }
 
         // POST api/<ValuesController>
+        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpPost]
-        public async Task Post(int id, string cinematography)
+        public async Task<IResult> Post([FromForm] int id,[FromForm] string cinematography)
         {
-            var _ = Enum.TryParse<Cinematography>(cinematography, true, out Cinematography result);
-            await myFavoriService.AddToFavoritesAsync(id, result, this.User);
+            if (Enum.TryParse(cinematography, true, out Cinematography result))
+            {
+                try
+                {
+                    await myFavoriService.AddToFavoritesAsync(id, result, this.User);
+                    return Results.Created();
+                }
+                catch (Exception)
+                {
+                    return Results.Problem(statusCode: 500);
+                }
+            }
+            else
+                return Results.NotFound(cinematography);
         }
 
-        // PUT api/<ValuesController>/5
-        [HttpPut("{id}")]
-        public void Put(int id, [FromBody] string value)
-        {
-        }
+        //// PUT api/<ValuesController>/5
+        //[HttpPut("{id}")]
+        //public void Put(int id, [FromBody] string value)
+        //{
+        //}
 
         // DELETE api/<ValuesController>/5
-        [HttpDelete("{cinematography}/{id}")]
-        public async Task Delete(int id, string cinematography)
+        [ProducesResponseType(StatusCodes.Status202Accepted)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [HttpDelete]
+        public async Task<IResult> Delete([FromForm] int id, [FromForm] string cinematography)
         {
-            var _ = Enum.TryParse<Cinematography>(cinematography, true, out Cinematography result);
-            await myFavoriService.RemoveFromFavoritesAsync(id, result, this.User);
+            if (Enum.TryParse(cinematography, true, out Cinematography result))
+            {
+                try
+                {
+                    await myFavoriService.RemoveFromFavoritesAsync(id, result, this.User);
+                    return Results.Accepted();
+                }
+                catch (Exception)
+                {
+                    return Results.Problem(statusCode: 500);
+                }
+            }
+            else
+                return Results.NotFound(cinematography);
         }
     }
 }
